Updated with better code
There was an error in the logic where i was trying to find only the greatest prime number of a large number. That view would ignore the other large prime factors and would only show the greatest prime factor. I have included a list traversal which solves this. It is however not very efficient and is able to find prime factors of numbers with up to 14 digits in about 5 seconds. 
I'll look into it and see if there is another more efficient way to do this.
Thanks